# api/docs/swagger.yaml
definitions:
...
  schema.AcademicSession:
    properties:
      end_date:
        type: string
      name:
        type: string
      start_date:
        type: string
    type: object
...
  schema.Assistant:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      role:
        type: string
    type: object
...
  schema.BucketInfo:
    properties:
      contents:
        items:
          type: string
        type: array
      created:
        type: string
      name:
        type: string
      updated:
        type: string
    type: object
...
  schema.CollectionRequirement:
    properties:
      name:
        type: string
      options:
        items: {}
        type: array
      required:
        type: integer
      type:
        type: string
    type: object
...
  schema.Location:
    properties:
      building:
        type: string
      map_uri:
        type: string
      room:
        type: string
    type: object
...
  schema.Meeting:
    properties:
      end_date:
        type: string
      end_time:
        type: string
      location:
        $ref: '#/definitions/schema.Location'
      meeting_days:
        items:
          type: string
        type: array
      modality:
        type: string
      start_date:
        type: string
      start_time:
        type: string
    type: object
...
  schema.ObjectInfo:
    properties:
      bucket:
        type: string
      content_encoding:
        type: string
      content_type:
        type: string
      created:
        type: string
      md5:
        items:
          type: integer
        type: array
      media_link:
        type: string
      name:
        type: string
      size:
        type: integer
      updated:
        type: string
    type: object
...
  schema.Section:
    properties:
      _id:
        type: string
      academic_session:
        $ref: '#/definitions/schema.AcademicSession'
      attributes: {}
      core_flags:
        items:
          type: string
        type: array
      course_reference:
        type: string
      grade_distribution:
        items:
          type: integer
        type: array
      instruction_mode:
        type: string
      internal_class_number:
        type: string
      meetings:
        items:
          $ref: '#/definitions/schema.Meeting'
        type: array
      professors:
        items:
          type: string
        type: array
      section_corequisites:
        $ref: '#/definitions/schema.CollectionRequirement'
      section_number:
        type: string
      syllabus_uri:
        type: string
      teaching_assistants:
        items:
          $ref: '#/definitions/schema.Assistant'
        type: array
    type: object
...
host: api.utdnebula.com
info:
  contact: {}
  description: The developer Nebula Labs API for access to pertinent UT Dallas data
  title: dev-nebula-api
  version: 1.0.0
paths:
...
  /section:
    get:
      description: '"Returns paginated list of sections matching the query''s string-typed
        key-value pairs. See offset for more details on pagination."'
      operationId: sectionSearch
      parameters:
      - description: The starting position of the current page of sections (e.g. For
          starting at the 17th professor, offset=16).
        in: query
        name: offset
        type: number
      - description: The section's official number
        in: query
        name: section_number
        type: string
      - description: An id that points to the course in MongoDB that this section
          is an instantiation of
        in: query
        name: course_reference
        type: string
      - description: The name of the academic session of the section
        in: query
        name: academic_session.name
        type: string
      - description: The date of classes starting for the section
        in: query
        name: academic_session.start_date
        type: string
      - description: The date of classes ending for the section
        in: query
        name: academic_session.end_date
        type: string
      - description: One of the professors teaching the section
        in: query
        name: professors
        type: string
      - description: The first name of one of the teaching assistants of the section
        in: query
        name: teaching_assistants.first_name
        type: string
      - description: The last name of one of the teaching assistants of the section
        in: query
        name: teaching_assistants.last_name
        type: string
      - description: The role of one of the teaching assistants of the section
        in: query
        name: teaching_assistants.role
        type: string
      - description: The email of one of the teaching assistants of the section
        in: query
        name: teaching_assistants.email
        type: string
      - description: The internal (university) number used to reference this section
        in: query
        name: internal_class_number
        type: string
      - description: The instruction modality for this section
        in: query
        name: instruction_mode
        type: string
      - description: The start date of one of the section's meetings
        in: query
        name: meetings.start_date
        type: string
      - description: The end date of one of the section's meetings
        in: query
        name: meetings.end_date
        type: string
      - description: One of the days that one of the section's meetings
        in: query
        name: meetings.meeting_days
        type: string
      - description: The time one of the section's meetings starts
        in: query
        name: meetings.start_time
        type: string
      - description: The time one of the section's meetings ends
        in: query
        name: meetings.end_time
        type: string
      - description: The modality of one of the section's meetings
        in: query
        name: meetings.modality
        type: string
      - description: The building of one of the section's meetings
        in: query
        name: meetings.location.building
        type: string
      - description: The room of one of the section's meetings
        in: query
        name: meetings.location.room
        type: string
      - description: A hyperlink to the UTD room locator of one of the section's meetings
        in: query
        name: meetings.location.map_uri
        type: string
      - description: One of core requirement codes this section fulfills
        in: query
        name: core_flags
        type: string
      - description: A link to the syllabus on the web
        in: query
        name: syllabus_uri
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: A list of sections
          schema:
            items:
              $ref: '#/definitions/schema.Section'
            type: array
  /section/{id}:
    get:
      description: '"Returns the section with given ID"'
      operationId: sectionById
      parameters:
      - description: ID of the section to get
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: A section
          schema:
            $ref: '#/definitions/schema.Section'
...
  /storage/{bucket}:
    delete:
      description: '"Delete a bucket. This route is restricted to only Nebula Labs
        internal Projects."'
      operationId: deleteBucket
      parameters:
      - description: Name of the bucket
        in: path
        name: bucket
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - storage_key: []
    get:
      description: '"Get info on a bucket. This route is restricted to only Nebula
        Labs internal Projects."'
      operationId: bucketInfo
      parameters:
      - description: Name of the bucket
        in: path
        name: bucket
        required: true
        type: string
      responses:
        "200":
          description: The bucket's info
          schema:
            $ref: '#/definitions/schema.BucketInfo'
      security:
      - storage_key: []
  /storage/{bucket}/{objectID}:
    delete:
      description: '"Delete an object from a bucket. This route is restricted to only
        Nebula Labs internal Projects."'
      operationId: deleteObject
      parameters:
      - description: Name of the bucket
        in: path
        name: bucket
        required: true
        type: string
      - description: ID of the object
        in: path
        name: objectID
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - storage_key: []
    get:
      description: '"Get info on an object in a bucket. This route is restricted to
        only Nebula Labs internal Projects."'
      operationId: objectInfo
      parameters:
      - description: Name of the bucket
        in: path
        name: bucket
        required: true
        type: string
      - description: ID of the object
        in: path
        name: objectID
        required: true
        type: string
      responses:
        "200":
          description: The object's info
          schema:
            $ref: '#/definitions/schema.ObjectInfo'
      security:
      - storage_key: []
    post:
      description: '"Upload an object to a bucket. This route is restricted to only
        Nebula Labs internal Projects."'
      operationId: postObject
      parameters:
      - description: Name of the bucket
        in: path
        name: bucket
        required: true
        type: string
      - description: ID of the object
        in: path
        name: objectID
        required: true
        type: string
      - description: The data to upload
        in: formData
        name: data
        required: true
        type: file
      responses:
        "200":
          description: The object's info
          schema:
            $ref: '#/definitions/schema.ObjectInfo'
      security:
      - storage_key: []
...
schemes:
- http
- https
security:
- api_key: []
securityDefinitions:
  api_key:
    in: header
    name: x-api-key
    type: apiKey
  storage_key:
    in: header
    name: x-storage-key
    type: apiKey
swagger: "2.0"
...